Expose vello's default `wgpu` feature with an optional passthrough feature
https://github.com/linebender/vello/pull/359 made `wgpu` an optional
but default dependency/feature, but any crate using `vello_svg` will be
forced to opt in to `wgpu` due to not disabling the default here.

Since `vello_svg` does not need any rendering backend at all, this
feature should be disabled by default.  As the `vello` crate is
reexported from the root, a `wgpu` passthrough feature is provided so
that users can turn it back on without explicitly depending on `vello`
in their crate dependencies.